refactor(archive): peel diff and rename into views/archive/ helpers
Follows the stateless ArchiveExtract/ArchiveMount collaborator pattern: helpers
hold only self.tab; transient state (_t, is_editing, renamed_archive_original_name)
stays on the tab. Pure structural move, no behavior change.
